requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-Testing==0.4
ipdb==0.8
Flask-Admin==1.0.7
Flask-Babel==0.9
Flask-KVSession==0.4
Flask-Login==0.2.7
Flask-Migrate==0.1.4
Flask-Script==0.6.2
Flask-WTF==0.9.3
Flask-FeatureFlags==0.3
Flask-Cache==0.12
Flask-Mail==0.9.0
MySQL-python==1.2.5
SQLAlchemy-Utils==0.22.1
tablib
